Improve codegen (#1015)
* Add support for spitting comments in the generated proto code.

* Add support for passing comments to the generated Go interface file


Former-commit-id: 87d0816b70392202c4978aa5fe49e7ab4ca82913

+++ mixer/tools/codegen/pkg/interfacegen/generator_test.go
@@ -63,11 +63,13 @@ func TestGenerator_Generate(t *testing.T) {
 			}
 
 			defer func() {
-				if removeErr := os.Remove(oIntface.Name()); removeErr != nil {
-					t.Logf("Could not remove temporary file %s: %v", oIntface.Name(), removeErr)
-				}
-				if removeErr := os.Remove(oTmpl.Name()); removeErr != nil {
-					t.Logf("Could not remove temporary file %s: %v", oTmpl.Name(), removeErr)
+				if !t.Failed() {
+					if removeErr := os.Remove(oIntface.Name()); removeErr != nil {
+						t.Logf("Could not remove temporary file %s: %v", oIntface.Name(), removeErr)
+					}
+					if removeErr := os.Remove(oTmpl.Name()); removeErr != nil {
+						t.Logf("Could not remove temporary file %s: %v", oTmpl.Name(), removeErr)
+					}
 				}
 			}()
